+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 #### New Features
 
-- Added support for creating permanent UDFs/UDTFs with `DataFrame/Series/GroupBy.apply`, `map`, and `transform` by passing the `snowflake_udf_params` keyword argument. See documentation for details.
+- Added support for creating permanent and immutable UDFs/UDTFs with `DataFrame/Series/GroupBy.apply`, `map`, and `transform` by passing the `snowflake_udf_params` keyword argument. See documentation for details.
 
 ## 1.35.0 (YYYY-MM-DD)